+++ packages/addon-dev/src/rollup-declarations.ts
@@ -2,19 +2,68 @@ import type { Plugin } from 'rollup';
 import execa from 'execa';
 import walkSync from 'walk-sync';
 import { readFile, writeFile } from 'fs/promises';
-import { existsSync } from 'fs';
+import { existsSync, readFileSync } from 'fs';
+import { packageUp } from 'package-up';
+
+let glint1 = 'glint --declaration';
+let glint2 = 'ember-tsc --declaration';
 
 export default function rollupDeclarationsPlugin(
   declarationsDir: string,
-  command = 'glint --declaration'
+  /**
+   * The command to use to generate types.
+   * Defaults to:
+   * - glint --declaration     # for glint v1
+   * - ember-tsc --declaration # for glint v2
+   */
+  command?: string
 ): Plugin {
   let glintPromise: Promise<void>;
+
+  let commandToRun = command;
+
+  async function determineCommand() {
+    if (commandToRun) return;
+
+    let manifestPath = await packageUp();
+    if (!manifestPath) {
+      /**
+       * Historical default is to use glint v1
+       */
+      commandToRun = glint1;
+      return;
+    }
+
+    let manifestBuffer = readFileSync(manifestPath);
+    let manifest = JSON.parse(manifestBuffer.toString());
+    let deps = {
+      ...manifest.devDependencies,
+      ...manifest.dependencies,
+    };
+
+    if (deps['@glint/ember-tsc']) {
+      commandToRun = glint2;
+      return;
+    }
+    if (deps['@glint/core']) {
+      commandToRun = glint1;
+      return;
+    }
+
+    throw new Error(
+      `Cannot use addon.declarations() plugin without glint present or an explicit command set as the second parameter. e.g.: addon.declarations('declarations', 'tsc --declaration')`
+    );
+  }
 
   return {
     name: 'declarations',
     buildStart() {
       const runGlint = async () => {
-        let { exitCode, escapedCommand } = await execa.command(command, {
+        await determineCommand();
+
+        if (!commandToRun) return;
+
+        let { exitCode, escapedCommand } = await execa.command(commandToRun, {
           // using stdio: inherit is the only way to retain color output from the
           // underlying tsc process.
           // However, the viewer of the error will not know which plugin it comes from.