+++ packages/reverse-exports/tests/smoke.ts
@@ -1,0 +1,14 @@
+import { externalName } from '../src';
+import assert from 'assert';
+
+const actual = externalName(
+  {
+    name: 'my-addon',
+    version: '1.1.0',
+    exports: {
+      './*': './dist/*.js',
+    },
+  },
+  './dist/foo.js'
+);
+assert.strictEqual(actual, 'my-addon/foo');