+++ packages/vite/src/build.ts
@@ -57,6 +57,7 @@ export function compatPrebuild(): Plugin {
   let viteCommand: string | undefined;
   let viteMode: string | undefined;
   let resolvableExtensions: string[] | undefined;
+  let buildPromise: Promise<void> | undefined;
 
   return {
     name: 'embroider-builder',
@@ -73,13 +74,19 @@ export function compatPrebuild(): Plugin {
     // then our resolver would get initialized without the meta data needed from compat-prebuild.
     // See https://github.com/embroider-build/embroider/issues/2490
     async options() {
+      if (buildPromise) {
+        await buildPromise;
+        return;
+      }
       if (!viteCommand) {
         throw new Error(`bug: embroider compatPrebuild did not detect Vite's command`);
       }
       if (!viteMode) {
         throw new Error(`bug: embroider compatPrebuild did not detect Vite's mode`);
       }
-      await emberBuild(viteCommand, viteMode, resolvableExtensions);
+
+      buildPromise = emberBuild(viteCommand, viteMode, resolvableExtensions);
+      await buildPromise;
     },
   };
 }