+++ tests/fixtures/macro-sample-addon/ember-cli-build.js
@@ -15,8 +15,5 @@ module.exports = function(defaults) {
     behave. You most likely want to be modifying `./index.js` or app's build file
   */
 
-  return maybeEmbroider(app, {
-    useAddonAppBoot: false,
-    useAddonConfigModule: false,
-  });
+  return app.toTree();
 };